refactor(BA-6330): import BraceStyleAdapter from ai.backend.logging in accelerator plugins (#11988)
Backported-from: main (26.4)
Backported-to: 26.4
Backport-of: 11988
