+++ changes/11988.enhance.md
@@ -1,0 +1,1 @@
+Import BraceStyleAdapter from ai.backend.logging instead of the deprecated ai.backend.common.logging shim in accelerator plugins
