+++ src/ai/backend/accelerator/tenstorrent/n300/plugin.py
@@ -41,7 +41,6 @@ from ai.backend.agent.stats import (
     StatContext,
 )
 from ai.backend.agent.types import Container, MountInfo
-from ai.backend.common.logging import BraceStyleAdapter
 from ai.backend.common.types import (
     AcceleratorMetadata,
     BinarySize,
@@ -52,6 +51,7 @@ from ai.backend.common.types import (
     SlotName,
     SlotTypes,
 )
+from ai.backend.logging import BraceStyleAdapter
 
 PREFIX = "tt-n300"
 VALID_CARD_TYPE = "n300"
